feat(onboarding): refactored package structure
