+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools">
+    <uses-permission android:name="android.permission.INTERNET" />
+
 
     <application
         android:allowBackup="true"
@@ -13,7 +15,7 @@
         android:theme="@style/Theme.Montra"
         tools:targetApi="31">
         <activity
-            android:name=".splashscreen.SplashScreenActivity"
+            android:name=".presentation.splashscreen.SplashScreenActivity"
             android:exported="true"
             android:theme="@style/Theme.Onboarding">
             <meta-data
@@ -28,7 +30,7 @@
 
         </activity>
         <activity
-            android:name=".onboarding.OnboardingActivity"
+            android:name=".presentation.onboarding.OnboardingActivity"
             android:exported="false"
             android:windowSoftInputMode="adjustResize"
             android:theme="@style/Theme.Montra">
