+++ app/src/main/java/com/app/montra/presentation/PinEntryEditText.kt
@@ -1,4 +1,4 @@
-package com.app.montra.views
+package com.app.montra.presentation
 
 import android.content.Context
 import android.content.res.ColorStateList
