+++ app/src/main/java/com/app/montra/presentation/onboarding/EmailSentFragment.kt
@@ -1,4 +1,4 @@
-package com.app.montra.onboarding
+package com.app.montra.presentation.onboarding
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,7 +7,6 @@ import android.view.ViewGroup
 import androidx.navigation.findNavController
 import com.app.montra.R
 import com.app.montra.databinding.FragmentEmailSentBinding
-import com.app.montra.databinding.FragmentForgotPasswordBinding
 
 class EmailSentFragment : BaseFragment() {
 
@@ -34,7 +33,8 @@ class EmailSentFragment : BaseFragment() {
 
     override fun onClick() {
         binding.btnBackToEmail.btnGenericLarge.setOnClickListener {
-            val action = EmailSentFragmentDirections.actionEmailSentFragmentToLoginFragment()
+            val action =
+                com.app.montra.onboarding.EmailSentFragmentDirections.actionEmailSentFragmentToLoginFragment()
             requireActivity().findNavController(R.id.nav_host_fragment).navigate(action)
         }
     }
