+++ app/src/main/java/com/app/montra/presentation/onboarding/SetUpAccountFragment.kt
@@ -1,11 +1,10 @@
-package com.app.montra.onboarding
+package com.app.montra.presentation.onboarding
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import com.app.montra.R
-import com.app.montra.databinding.FragmentOnboardingBinding
 import com.app.montra.databinding.FragmentSetUpAccountBinding
 
 class SetUpAccountFragment : BaseFragment() {
