+++ app/src/main/java/com/app/montra/presentation/onboarding/SignUpFragment.kt
@@ -1,4 +1,4 @@
-package com.app.montra.onboarding
+package com.app.montra.presentation.onboarding
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,12 +7,24 @@ import android.view.ViewGroup
 import androidx.navigation.findNavController
 import com.app.montra.R
 import com.app.montra.databinding.FragmentSignUpBinding
+import com.auth0.android.Auth0
+
 
 class SignUpFragment : BaseFragment() {
 
     private lateinit var _binding: FragmentSignUpBinding
     private val binding get() = _binding
+    private lateinit var account: Auth0
 
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        // Set up the account object with the Auth0 application details
+        account = Auth0(
+            getString(R.string.auth0_client_id),
+            getString(R.string.auth0_domain)
+        )
+    }
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
@@ -28,7 +40,8 @@ class SignUpFragment : BaseFragment() {
         }
 
         binding.btnOnboardingSignup.btnGenericLarge.setOnClickListener {
-            val action = SignUpFragmentDirections.actionSignupFragmentToVerificationFragment()
+            val action =
+                com.app.montra.onboarding.SignUpFragmentDirections.actionSignupFragmentToVerificationFragment()
             requireActivity().findNavController(R.id.nav_host_fragment).navigate(action)
         }
     }
