+++ app/src/main/java/com/app/montra/presentation/onboarding/VerificationFragment.kt
@@ -1,12 +1,10 @@
-package com.app.montra.onboarding
+package com.app.montra.presentation.onboarding
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import com.app.montra.R
-import com.app.montra.databinding.FragmentResetPasswordBinding
 import com.app.montra.databinding.FragmentVerificationBinding
 
 class VerificationFragment : BaseFragment() {
