+++ app/src/main/java/com/app/montra/presentation/splashscreen/SplashScreenActivity.kt
@@ -1,11 +1,11 @@
-package com.app.montra.splashscreen
+package com.app.montra.presentation.splashscreen
 
 import android.annotation.SuppressLint
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
-import com.app.montra.onboarding.OnboardingActivity
+import com.app.montra.presentation.onboarding.OnboardingActivity
 import com.app.montra.R
 
 @SuppressLint("CustomSplashScreen")
