+++ app/src/main/res/layout/fragment_add_new_account.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/violet_100"
-    tools:context=".onboarding.AddNewAccountFragment">
+    tools:context=".presentation.onboarding.AddNewAccountFragment">
 
     <include
         android:id="@+id/toolbar"
