+++ app/src/main/res/layout/fragment_email_sent.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".onboarding.EmailSentFragment">
+    tools:context=".presentation.onboarding.EmailSentFragment">
 
     <androidx.appcompat.widget.AppCompatImageView
         android:id="@+id/email_flying_image"
