+++ app/src/main/res/layout/fragment_onboarding.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".onboarding.OnBoardingFragment">
+    tools:context=".presentation.onboarding.OnBoardingFragment">
 
     <androidx.viewpager.widget.ViewPager
         android:id="@+id/onboarding_viewpager"
