+++ app/src/main/res/layout/fragment_pin_entry.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/violet_100"
-    tools:context=".onboarding.PinEntryFragment">
+    tools:context=".presentation.onboarding.PinEntryFragment">
 
 <!--    &lt;!&ndash; TODO: Update blank fragment layout &ndash;&gt;-->
     <androidx.appcompat.widget.AppCompatTextView
@@ -30,7 +30,7 @@
         app:layout_constraintTop_toBottomOf="@+id/pin_entry_heading"
         android:layout_height="wrap_content">
 
-        <com.app.montra.views.PinEntryEditText
+        <com.app.montra.presentation.PinEntryEditText
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:backgroundTint="@color/light_80"
