+++ app/src/main/res/layout/fragment_reset_password.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".onboarding.ResetPasswordFragment">
+    tools:context=".presentation.onboarding.ResetPasswordFragment">
 
     <include
         android:id="@+id/toolbar"
