+++ app/src/main/res/layout/fragment_set_up_account.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_marginHorizontal="@dimen/margin16"
-    tools:context=".onboarding.SetUpAccountFragment">
+    tools:context=".presentation.onboarding.SetUpAccountFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
