+++ app/src/main/res/layout/fragment_setup_finished.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".onboarding.SetupFinishedFragment">
+    tools:context=".presentation.onboarding.SetupFinishedFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <androidx.appcompat.widget.AppCompatImageView
