+++ app/src/main/res/layout/fragment_sign_up.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".onboarding.SignUpFragment">
+    tools:context=".presentation.onboarding.SignUpFragment">
 
     <include
         android:id="@+id/toolbar"
